Release the GVL while calls to C code
FFI defaults to not release the GVL while calls to C code
(in contrast to Fiddle). This blocks other Ruby threads for
the time of the call. It also slows simultanous communication
down, when running over several TTY's in parallel.

The negative performance impact of releasing the GVL is so small,
that it doesn't matter compared to the slowness of RS232.
However the positive performance impact when running multiple
threads is huge: In my case with 16 devices releasing GVL
drops the time for one particular operation from 160 to 23 seconds.